fix : works but a bit slow to my taste

+++ finalCalc/src/main/java/fr/cybercicco/deckentities/Player.java
@@ -7,7 +7,7 @@ import java.util.List;
 
 public class Player {
     public List<Card> hand = new ArrayList<>();
-    public long strength;
+    public double strength;
     public List<Card> combination = new ArrayList<>();
 
     public void setHand(Deck deck, String cards){
